HBX-1673: Move class 'org.hibernate.tool.hbmlint.detector.RelationalModelDetector' to 'org.hibernate.tool.internal.export.lint.RelationalModelDetector'
